ApiServer/ML_requirements.txt
absl-py==0.9.0
asgiref==3.4.1
asn1crypto==0.24.0
astor==0.8.1
attrs==17.4.0
Automat==0.6.0
bleach==2.1.2
blessed==1.18.0
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
cloud-init==19.4
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
decorator==4.1.2
distro-info===0.18ubuntu0.18.04.1
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
entrypoints==0.2.3.post1
gast==0.3.3
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
html5lib==0.999999999
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
ipykernel==4.8.2
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
Jinja2==2.10
joblib==1.0.1
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-core==4.4.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
Markdown==3.2.1
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
netifaces==0.10.4
notebook==5.2.2
numpy==1.18.1
oauthlib==2.0.6
opencv-contrib-python==4.5.2.52
opencv-python==4.5.2.52
PAM==0.4.2
pandocfilters==1.4.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==8.2.0
prompt-toolkit==1.0.15
protobuf==3.11.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
Pygments==2.2.0
pygobject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2021.1
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.2
requests==2.18.4
requests-unixsocket==0.1.5
scikit-learn==0.24.2
scipy==1.4.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
service-identity==16.0.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
sqlparse==0.4.1
ssh-import-id==5.7
systemd-python==234
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.7
testpath==0.3.1
threadpoolctl==2.1.0
tornado==4.5.3
traitlets==4.3.2
Twisted==17.9.0
typing-extensions==3.10.0.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5
Werkzeug==1.0.0
wrapt==1.12.0
xkit==0.0.0
zope.interface==4.3.2
